Now ready for batch processing.

+++ src/kernel.h
@@ -3,7 +3,7 @@
 class Kernel
 {
 public:
-    Kernel(int w, int h, float *k, int n = 1) { width = w; height=h; this->k = k; norm = n; }
+    Kernel(int w, int h, float *k, float n = 1) { width = w; height=h; this->k = k; norm = n; }
     ~Kernel();
     int get_height() { return height; }
     int get_width() { return width; }
